Move SingletonTask<T> to SpecializedTasks.cs and add comments.
