+++ src/Workspaces/Core/Portable/Utilities/TaskExtensions.cs
@@ -363,7 +363,7 @@ namespace Roslyn.Utilities
 
             if (cancellationToken.IsCancellationRequested)
             {
-                return SingletonTask<T>.CanceledTask;
+                return SpecializedTasks.FromCancellationToken<T>(cancellationToken);
             }
 
             return WithCancellationSlow(task, cancellationToken);
@@ -390,7 +390,7 @@ namespace Roslyn.Utilities
 
             if (cancellationToken.IsCancellationRequested)
             {
-                return SingletonTask<int>.CanceledTask;
+                return SpecializedTasks.FromCancellationToken<int>(cancellationToken);
             }
 
             return WithCancellationSlow(task, cancellationToken);
@@ -466,28 +466,6 @@ namespace Roslyn.Utilities
             // > ~67s     // switch to thread 67
             // > !dso     // dump stack objects
             FatalError.Report(exception);
-        }
-
-        /// <summary>
-        /// Wraps a Task{T} that has already been canceled.
-        /// </summary>
-        /// <typeparam name="T">The type of value that might have been returned by the task except for its cancellation.</typeparam>
-        private static class SingletonTask<T>
-        {
-            /// <summary>
-            /// A task that is already canceled.
-            /// </summary>
-            internal static readonly Task<T> CanceledTask = CreateCanceledTask();
-
-            /// <summary>
-            /// Creates a canceled task.
-            /// </summary>
-            private static Task<T> CreateCanceledTask()
-            {
-                var tcs = new TaskCompletionSource<T>();
-                tcs.SetCanceled();
-                return tcs.Task;
-            }
         }
     }
 }
